Move common functionality to AnnotationChooser
