+++ src/main/kotlin/io/github/trubitsyn/visiblefortesting/AnnotateMethodFromTestIntention.kt
@@ -22,9 +22,7 @@ import com.intellij.ide.projectView.impl.ProjectRootsUtil
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
 import com.intellij.psi.*
-import com.intellij.psi.search.GlobalSearchScope
 import com.intellij.psi.util.PsiUtil
-import com.intellij.ui.popup.list.ListPopupImpl
 import com.intellij.util.IncorrectOperationException
 import org.jetbrains.annotations.NonNls
 
@@ -63,26 +61,11 @@ class AnnotateMethodFromTestIntention : BaseElementAtCaretIntentionAction() {
     @Throws(IncorrectOperationException::class)
     override fun invoke(project: Project, editor: Editor, element: PsiElement) {
         val call = element.parent.parent as PsiMethodCallExpression
+        val method = call.resolveMethod() ?: return
 
-        call.resolveMethod()?.let {
-            val annotations = Annotations.getAvailable(project)
-
-            if (annotations.size == 1) {
-                if (!AnnotationApplier.isAnnotated(it, annotations[0])) {
-                    AnnotationApplier.addAnnotation(it, annotations[0])
-                }
-            } else {
-                val facade = JavaPsiFacade.getInstance(project)
-                val scope = GlobalSearchScope.allScope(project)
-                val psiClasses = annotations.map { facade.findClass(it.qualifiedName, scope) }
-
-                val importDialog = ChooseClassDialog(psiClasses, project, { psiClass ->
-                    AnnotationApplier.addAnnotation(it, annotations.first { it.qualifiedName == psiClass.qualifiedName})
-                })
-
-                ListPopupImpl(importDialog).showInBestPositionFor(editor)
-            }
-        }
+        AnnotationChooser.choose(project, editor, {
+            AnnotationApplier.addAnnotation(method, it)
+        })
     }
 
     override fun startInWriteAction() = true
